global: migrate to (api_)finalize_app
* record_once is not deprecated but since the before_(app)_first_request
  decorators are deprecated this has been moved also to the new
  invenio-base entrypoints

* the blueprint_ui_init and blueprint_api_init have been removed,
  because they are only used as the blueprints to run record_once
