+++ setup.cfg
@@ -49,12 +49,12 @@ sqlite =
 [options.entry_points]
 invenio_base.api_apps =
     invenio_userprofiles = invenio_userprofiles:InvenioUserProfiles
-invenio_base.api_blueprints =
-    invenio_userprofiles = invenio_userprofiles.views:blueprint_api_init
+invenio_base.api_finalize_app =
+    invenio_userprofiles = invenio_userprofiles.ext:api_finalize_app
 invenio_base.apps =
     invenio_userprofiles = invenio_userprofiles:InvenioUserProfiles
-invenio_base.blueprints =
-    invenio_userprofiles = invenio_userprofiles.views:blueprint_ui_init
+invenio_base.finalize_app =
+    invenio_userprofiles = invenio_userprofiles.ext:finalize_app
 invenio_db.alembic =
     invenio_userprofiles = invenio_userprofiles:alembic
 invenio_i18n.translations =
